@@ -1,6 +1,8 @@
 module istio.io/istio.io
 
-go 1.21
+go 1.22
+
+toolchain go1.22.1
 
 // https://github.com/containerd/containerd/issues/5781
 exclude k8s.io/kubernetes v1.13.0
