@@ -257,10 +257,11 @@ TEST(PosixThreadTest, PThreadId) {
   auto thread =
       thread_factory->createThread([&]() { thread_id = thread_factory->currentPthreadId(); },
                                    /* options= */ absl::nullopt, /* crash_on_failure= */ false);
+  auto threadId = thread->pthreadId();
   thread->join();
 
-  EXPECT_EQ(thread->pthreadId(), thread_id);
-  EXPECT_NE(thread->pthreadId(), thread_factory->currentThreadId());
+  EXPECT_EQ(threadId, thread_id);
+  EXPECT_NE(threadId, thread_factory->currentThreadId());
 }
 
 TEST(PosixThreadTest, Joinable) {
